// src/types.ts
export interface Extensable {
    [key: `x-${string}`]: any
}
// ...
export interface OpenAPI3 extends Extensable {
    /** REQUIRED. This string MUST be the version number of the OpenAPI Specification that the OpenAPI document uses. The openapi field SHOULD be used by tooling to interpret the OpenAPI document. This is not related to the API info.version string. */
    openapi: string
    /** REQUIRED. Provides metadata about the API. The metadata MAY be used by tooling as required. */
    info: InfoObject // required
    /** The default value for the $schema keyword within Schema Objects contained within this OAS document. This MUST be in the form of a URI. */
    jsonSchemaDialect?: string
    /** An array of Server Objects, which provide connectivity information to a target server. If the servers property is not provided, or is an empty array, the default value would be a Server Object with a url value of /. */
    servers?: ServerObject[]
    /** The available paths and operations for the API. */
    paths?: PathsObject
    /** The incoming webhooks that MAY be received as part of this API and that the API consumer MAY choose to implement. Closely related to the callbacks feature, this section describes requests initiated other than by an API call, for example by an out of band registration. The key name is a unique string to refer to each webhook, while the (optionally referenced) Path Item Object describes a request that may be initiated by the API provider and the expected responses. An example is available. */
    webhooks?: { [id: string]: PathItemObject | ReferenceObject }
    /** An element to hold various schemas for the document. */
    components?: ComponentsObject
    /** A declaration of which security mechanisms can be used across the API. The list of values includes alternative security requirement objects that can be used. Only one of the security requirement objects need to be satisfied to authorize a request. Individual operations can override this definition. To make security optional, an empty security requirement ({}) can be included in the array. */
    security?: SecurityRequirementObject[]
    /** A list of tags used by the document with additional metadata. The order of the tags can be used to reflect on their order by the parsing tools. Not all tags that are used by the Operation Object must be declared. The tags that are not declared MAY be organized randomly or based on the tools’ logic. Each tag name in the list MUST be unique. */
    tags?: TagObject[]
    /** Additional external documentation. */
    externalDocs?: ExternalDocumentationObject
    $defs?: $defs
}

/**
 * [4.8.2] Info Object
 * The object provides metadata about the API.
 */
export interface InfoObject extends Extensable {
    /** REQUIRED. The title of the API. */
    title: string
    /** A short summary of the API. */
    summary?: string
    /** A description of the API. CommonMark syntax MAY be used for rich text representation. */
    description?: string
    /** A URL to the Terms of Service for the API. This MUST be in the form of a URL. */
    termsOfService?: string
    /** The contact information for the exposed API. */
    contact?: ContactObject
    /** The license information for the exposed API. */
    license?: LicenseObject
    /** REQUIRED. The version of the OpenAPI document (which is distinct from the OpenAPI Specification version or the API implementation version). */
    version: string
}

/**
 * [4.8.3] Contact Object
 * Contact information for the exposed API.
 */
export interface ContactObject extends Extensable {
    /** The identifying name of the contact person/organization. */
    name?: string
    /** The URL pointing to the contact information. This MUST be in the form of a URL. */
    url?: string
    /** The email address of the contact person/organization. This MUST be in the form of an email address. */
    email?: string
}

/**
 * [4.8.4] License object
 * License information for the exposed API.
 */
export interface LicenseObject extends Extensable {
    /** REQUIRED. The license name used for the API. */
    name: string
    /** An SPDX license expression for the API. The identifier field is mutually exclusive of the url field. */
    identifier: string
    /** A URL to the license used for the API. This MUST be in the form of a URL. The url field is mutually exclusive of the identifier field. */
    url: string
}

/**
 * [4.8.5] Server Object
 * An object representing a Server.
 */
export interface ServerObject extends Extensable {
    /** REQUIRED. A URL to the target host. This URL supports Server Variables and MAY be relative, to indicate that the host location is relative to the location where the OpenAPI document is being served. Variable substitutions will be made when a variable is named in {brackets}. */
    url: string
    /** An optional string describing the host designated by the URL. CommonMark syntax MAY be used for rich text representation. */
    description: string
    /** A map between a variable name and its value. The value is used for substitution in the server’s URL template. */
    variables?: { [name: string]: ServerVariableObject }
}

/**
 * [4.8.6] Server Variable Object
 * An object representing a Server Variable for server URL template substitution.
 */
export interface ServerVariableObject extends Extensable {
    /** An enumeration of string values to be used if the substitution options are from a limited set. The array MUST NOT be empty. */
    enum?: string[]
    /** REQUIRED. The default value to use for substitution, which SHALL be sent if an alternate value is not supplied. Note this behavior is different than the Schema Object’s treatment of default values, because in those cases parameter values are optional. If the enum is defined, the value MUST exist in the enum’s values. */
    default: string
    /** An optional description for the server variable. CommonMark syntax MAY be used for rich text representation. */
    description?: string
}

/**
 * [4.8.7] Components Object
 * Holds a set of reusable objects for different aspects of the OAS.
 */
export interface ComponentsObject extends Extensable {
    /** An object to hold reusable Schema Objects.*/
    schemas?: Record<string, SchemaObject>
    /** An object to hold reusable Response Objects. */
    responses?: Record<string, ResponseObject | ReferenceObject>
    /** An object to hold reusable Parameter Objects. */
    parameters?: Record<string, ParameterObject | ReferenceObject>
    /** An object to hold reusable Example Objects. */
    examples?: Record<string, ExampleObject | ReferenceObject>
    /** An object to hold reusable Request Body Objects. */
    requestBodies?: Record<string, RequestBodyObject | ReferenceObject>
    /** An object to hold reusable Header Objects. */
    headers?: Record<string, HeaderObject | ReferenceObject>
    /** An object to hold reusable Security Scheme Objects. */
    securitySchemes?: Record<string, SecuritySchemeObject | ReferenceObject>
    /** An object to hold reusable Link Objects. */
    links?: Record<string, LinkObject | ReferenceObject>
    /** An object to hold reusable Callback Objects. */
    callbacks?: Record<string, CallbackObject | ReferenceObject>
    /** An object to hold reusable Path Item Objects. */
    pathItems?: Record<string, PathItemObject | ReferenceObject>
}

/**
 * [4.8.8] Paths Object
 * Holds the relative paths to the individual endpoints and their operations. The path is appended to the URL from the Server Object in order to construct the full URL. The Paths MAY be empty, due to Access Control List (ACL) constraints.
 */
export interface PathsObject {
    [pathname: string]: PathItemObject | ReferenceObject // note: paths object does support $refs; the schema just defines it in a weird way
}
// ...
export interface PathItemObject extends Extensable {
    /** A definition of a GET operation on this path. */
    get?: OperationObject | ReferenceObject
    /** A definition of a PUT operation on this path. */
    put?: OperationObject | ReferenceObject
    /** A definition of a POST operation on this path. */
    post?: OperationObject | ReferenceObject
    /** A definition of a DELETE operation on this path. */
    delete?: OperationObject | ReferenceObject
    /** A definition of a OPTIONS operation on this path. */
    options?: OperationObject | ReferenceObject
    /** A definition of a HEAD operation on this path. */
    head?: OperationObject | ReferenceObject
    /** A definition of a PATCH operation on this path. */
    patch?: OperationObject | ReferenceObject
    /** A definition of a TRACE operation on this path. */
    trace?: OperationObject | ReferenceObject
    /** An alternative server array to service all operations in this path. */
    servers?: ServerObject[]
    /** A list of parameters that are applicable for all the operations described under this path. These parameters can be overridden at the operation level, but cannot be removed there. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location. The list can use the Reference Object to link to parameters that are defined at the OpenAPI Object’s components/parameters. */
    parameters?: (ParameterObject | ReferenceObject)[]
}
// ...
export interface ExternalDocumentationObject extends Extensable {
    /** A description of the target documentation. CommonMark syntax MAY be used for rich text representation. */
    description?: string
    /** REQUIRED. The URL for the target documentation. This MUST be in the form of a URL. */
    url: string
}
// ...
export type CallbackObject = Record<string, PathItemObject>
// ...
export type HeaderObject = Omit<ParameterObject, 'name' | 'in'>

/**
 * [4.8.22] Tag Object
 * Adds metadata to a single tag that is used by the Operation Object. It is not mandatory to have a Tag Object per tag defined in the Operation Object instances.
 */
export interface TagObject extends Extensable {
    /** REQUIRED. The name of the tag. */
    name: string
    /** A description for the tag. CommonMark syntax MAY be used for rich text representation. */
    description?: string
    /** Additional external documentation for this tag. */
    externalDocs?: ExternalDocumentationObject
}

/**
 * [4.8.23] Reference Object
 * A simple object to allow referencing other components in the OpenAPI document, internally and externally. The $ref string value contains a URI [RFC3986], which identifies the location of the value being referenced. See the rules for resolving Relative References.
 */
export interface ReferenceObject extends Extensable {
    /** REQUIRED. The reference identifier. This MUST be in the form of a URI. */
    $ref: string
    /** A short summary which by default SHOULD override that of the referenced component. If the referenced object-type does not allow a summary field, then this field has no effect. */
    summary?: string
    /** A description which by default SHOULD override that of the referenced component. CommonMark syntax MAY be used for rich text representation. If the referenced object-type does not allow a description field, then this field has no effect. */
    description?: string
}

/**
 * [4.8.24] Schema Object
 * The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is a superset of the JSON Schema Specification Draft 2020-12.
 */
export type SchemaObject = {
    /** The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is a superset of the JSON Schema Specification Draft 2020-12. */
    discriminator?: DiscriminatorObject
    /** MAY be used only on properties schemas. It has no effect on root schemas. Adds additional metadata to describe the XML representation of this property. */
    xml?: XMLObject
    /** Additional external documentation for this schema. */
    externalDocs?: ExternalDocumentationObject
    /** @deprecated */
    example?: any
    title?: string
    description?: string
    $comment?: string
    deprecated?: boolean
    readOnly?: boolean
    uniqueItems?: boolean
    writeOnly?: boolean
    enum?: unknown[]
    /** Use of this keyword is functionally equivalent to an "enum" (Section 6.1.2) with a single value. */
    const?: unknown
    default?: unknown
    format?: string
    /** @deprecated in 3.1 (still valid for 3.0) */
    nullable?: boolean
    oneOf?: (SchemaObject | ReferenceObject)[]
    allOf?: (SchemaObject | ReferenceObject)[]
    anyOf?: (SchemaObject | ReferenceObject)[]
    required?: string[]
    [key: `x-${string}`]: any
} & (
    | StringSubtype
    | NumberSubtype
    | IntegerSubtype
    | ArraySubtype
    | BooleanSubtype
    | NullSubtype
    | ObjectSubtype
    | {
          type: ('string' | 'number' | 'integer' | 'array' | 'boolean' | 'null' | 'object')[]
      }
    // eslint-disable-next-line @typescript-eslint/ban-types
    | {}
)
// ...
export type SecurityRequirementObject = Record<keyof ComponentsObject['securitySchemes'], string[]>

export type $defs = Record<string, SchemaObject>